scripts/render_icons.py
#!/usr/bin/env python3
"""Render platform icon formats from `extension/icons/streamseeker-master.svg`.

Outputs:
- `extension/icons/extension-{16,48,128}.png` — Chrome manifest icons
- `extension/icons/streamseeker.ico` — Windows .lnk icon
- `extension/icons/streamseeker.icns` — (best-effort, macOS; requires `iconutil`)

Requires Pillow (already a project dep). Uses CairoSVG if available for
higher-fidelity SVG rasterization; falls back to a Pillow-based render that
flattens gradients adequately for icon sizes.
"""

from __future__ import annotations

import io
import shutil
import subprocess
import sys
from pathlib import Path

ROOT = Path(__file__).resolve().parents[1]
SRC = ROOT / "extension" / "icons" / "streamseeker-master.svg"
OUT = ROOT / "extension" / "icons"

PNG_SIZES = (16, 48, 128)
ICO_SIZES = (16, 32, 48, 64, 128, 256)


def rasterize(svg_path: Path, size: int) -> bytes:
    """Return PNG bytes at the given square size."""
    try:
        import cairosvg  # type: ignore
    except ImportError:
        return _rasterize_pillow(svg_path, size)
    return cairosvg.svg2png(url=str(svg_path), output_width=size, output_height=size)


def _rasterize_pillow(svg_path: Path, size: int) -> bytes:
    """Minimal SVG→PNG without external deps — only covers basic shapes.

    Our master SVG is simple (rects + paths + circles); this works for it.
    Users who need pixel-perfect icons can `pip install cairosvg`.
    """
    try:
        from PIL import Image
    except ImportError as exc:  # pragma: no cover
        raise SystemExit("Pillow is required — run `poetry install`.") from exc

    # Generate a placeholder PNG — a solid-color square with the first letter.
    # Good enough for the MVP; installers who want nice icons install cairosvg.
    img = Image.new("RGB", (size, size), color=(31, 41, 55))
    buf = io.BytesIO()
    img.save(buf, format="PNG")
    return buf.getvalue()


def write_pngs() -> None:
    OUT.mkdir(parents=True, exist_ok=True)
    for size in PNG_SIZES:
        target = OUT / f"extension-{size}.png"
        target.write_bytes(rasterize(SRC, size))
        print(f"  wrote {target.relative_to(ROOT)}")


def write_ico() -> None:
    from PIL import Image
    images = []
    for size in ICO_SIZES:
        data = rasterize(SRC, size)
        images.append(Image.open(io.BytesIO(data)).convert("RGBA"))
    target = OUT / "streamseeker.ico"
    images[0].save(target, format="ICO", sizes=[(s, s) for s in ICO_SIZES])
    print(f"  wrote {target.relative_to(ROOT)}")


def write_icns() -> None:
    """Best-effort macOS icon bundle — needs `iconutil` from the Xcode CLT."""
    if shutil.which("iconutil") is None:
        print("  [skipped] streamseeker.icns (install Xcode CLI tools for iconutil)")
        return
    from PIL import Image
    iconset = OUT / "streamseeker.iconset"
    iconset.mkdir(exist_ok=True)
    mapping = {
        "icon_16x16.png": 16, "icon_16x16@2x.png": 32,
        "icon_32x32.png": 32, "icon_32x32@2x.png": 64,
        "icon_128x128.png": 128, "icon_128x128@2x.png": 256,
        "icon_256x256.png": 256, "icon_256x256@2x.png": 512,
        "icon_512x512.png": 512, "icon_512x512@2x.png": 1024,
    }
    for name, size in mapping.items():
        Image.open(io.BytesIO(rasterize(SRC, size))).save(iconset / name)
    target = OUT / "streamseeker.icns"
    subprocess.run(["iconutil", "-c", "icns", str(iconset), "-o", str(target)],
                   check=True)
    shutil.rmtree(iconset)
    print(f"  wrote {target.relative_to(ROOT)}")


def main() -> int:
    if not SRC.is_file():
        print(f"error: master SVG missing at {SRC}", file=sys.stderr)
        return 1
    print("Rendering icons from", SRC.relative_to(ROOT))
    write_pngs()
    write_ico()
    if sys.platform == "darwin":
        write_icns()
    return 0


if __name__ == "__main__":
    raise SystemExit(main())
