pyproject.toml
[tool.poetry]
name = "streamseeker"
version = "0.2.0"
description = "Helping package to download files from different streaming platforms"
authors = ["Dennis Fiedler <dennis.fiedler@uniprank.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/uniprank/streamseeker"
repository = "https://github.com/uniprank/streamseeker"
packages = [{ include = "streamseeker", from = "src" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
Issues = "https://github.com/uniprank/streamseeker/issues"

[tool.poetry.scripts]
streamseeker = "streamseeker.console.application:main"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
invoke = "^2.2.0"
requests = "^2.32.2"
selenium = "^4.15.2"
tqdm = "4.66.3"
tomli = "^2.0.2"
cleo = "^2.1.0"
numpy = "^2.1.2"
packaging = "^24.1"
fastapi = "^0.115.0"
uvicorn = { version = "^0.32.0", extras = ["standard"] }
Pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.0"
responses = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
